extracted ops sat data loader
